+++ desktop/electron/reportProcessing.ts
@@ -1,7 +1,10 @@
-import path from "path";
-import { Worker } from "worker_threads";
+// import { Worker } from "worker_threads";
 import { getProjectPath } from "./metadata";
 import audiences from "./mockData/audiences.json";
+import { createRequire } from 'node:module';
+
+// import Tinypool from 'tinypool'
+
 import type {
   Audience,
   ScriptData,
@@ -13,29 +16,29 @@ import script from "./mockData/script.json";
 import topics from "./mockData/topics.json";
 import visuals from "./mockData/visuals.json";
 import voiceovers from "./mockData/voiceovers.json";
+import Tinypool from 'tinypool'
 
+const require = createRequire(import.meta.url);
+let pool: Tinypool;
+
+try{
+  pool = new Tinypool({
+    filename: require.resolve('./tinypool.js'),
+  })
+}catch {
+  console.log(__dirname)
+  pool = new Tinypool({
+    filename:  require.resolve('../dist-electron/tinypool.js')
+  })
+}
 
 export async function textToAudio() {
   // return [generationId, audioUrl];
 }
 
-export function extractTextFromPDF(filePath: string): Promise<string> {
-  const worker = new Worker(path.resolve(__dirname, "./worker.js"), {});
-  worker.postMessage({ rpc: "setup", data: { path: getProjectPath() } });
-  worker.postMessage({ rpc: "extractTextFromPDF", data: { filePath } });
-
-  return new Promise((resolve, reject) => {
-    worker.on("message", ({ rpc, result }) => {
-      if (rpc === "extractTextFromPDF") {
-        resolve(result);
-      }
-    });
-    worker.on("error", reject);
-    worker.on("exit", (code) => {
-      if (code !== 0)
-        reject(new Error(`Worker stopped with exit code ${code}`));
-    });
-  });
+export async function extractTextFromPDF(filePath: string): Promise<string> {
+  const {text} = await (pool.run({filePath, projectPath: getProjectPath()},  { name: 'extractTextFromPDF' }) as Promise<{ text: string; images: ImageData[]; }>)
+  return text
 }
 
 export async function getScript(): Promise<ScriptData[]> {
